on product changing buying cost and selling price back to i from f

+++ models/product.rb
@@ -10,8 +10,8 @@ attr_accessor :product, :description, :stock_quantity, :buying_cost, :selling_pr
    @product = options['product']
    @description = options['description']
    @stock_quantity = options['stock_quantity'].to_i
-   @buying_cost = options['buying_cost'].to_f
-   @selling_price = options['selling_price'].to_f
+   @buying_cost = options['buying_cost'].to_i
+   @selling_price = options['selling_price'].to_i
    @supplier_id = options['supplier_id'].to_i
    # @comments = options['comments']
  end
@@ -25,7 +25,6 @@ attr_accessor :product, :description, :stock_quantity, :buying_cost, :selling_pr
      buying_cost,
      selling_price,
      supplier_id
-
      )
      VALUES
      ($1, $2, $3, $4, $5, $6)
